+++ src/Constants.cs
@@ -1,8 +1,0 @@
-﻿namespace TypeScriptDefinitionGenerator
-{
-    public static class Constants
-    {
-        public const string FileExtension = ".d.ts";
-        public static string[] SupportedSourceExtensions { get; } = { ".cs", ".vb" };
-    }
-}